[xiv] Increase delays for keyboard navigation
Until we narrow down the causes of the desync experienced
at times it seems prudent to play it safe for now. This will
affect Talan uses of items with many ingredients the most since
a recipe with 4-6 ingredients may add an additional second or two
to each craft, but that time can be made up by improvements to
material handling in general rather than the "IPC" layer.